base case of the recursiton, new file for checks
