+++ numalgsolve/polynomial.py
@@ -966,7 +966,7 @@ def is_power(poly_list, return_string = False):
         else:
             return 'MultiCheb'
     else:
-        print([type(p) == MultiPower for p in initial_poly_list])
+        print([type(p) == MultiPower for p in poly_list])
         raise ValueError('Bad polynomials in list')
 
 ############################################################################
